refactor(core): extract CollectionConfig and ColumnStore PK ops for NLOC compliance
- Extract CollectionConfig, CURRENT_SCHEMA_VERSION, and serde defaults
  from types.rs (581→372 lines) into collection_config.rs (223 lines)
- Extract 12 primary key CRUD methods from column_store/mod.rs (564→304
  lines) into primary_key_ops.rs (272 lines)
- Fix cargo fmt ordering in search/query/mod.rs and execution_paths.rs
- All files now under 500 NLOC production limit
- Zero duplication between extracted files (jscpd verified)
- Backward-compatible re-exports preserve all downstream imports
